requirements.txt
transformers==2.11.0
pytorch==1.4.0
pandas==1.0.3
xlrd==1.2.0
MeCab
